refactor: OptionDto 사용
