+++ src/test/java/gift/service/ProductServiceTest.java
@@ -3,9 +3,8 @@ package gift.service;
 import gift.domain.Category;
 import gift.domain.Option;
 import gift.domain.Product;
-import gift.dto.OptionDto;
 import gift.dto.ProductDto;
-import gift.dto.request.OptionRequest;
+import gift.dto.request.OptionDto;
 import gift.exception.ProductNotFoundException;
 import gift.repository.CategoryRepository;
 import gift.repository.ProductRepository;
@@ -86,7 +85,7 @@ class ProductServiceTest {
     void addProduct() throws Exception {
         //given
         Long categoryId = 1L;
-        ProductDto dto = new ProductDto("아이스티", 2500, "https://example.com", categoryId, List.of(new OptionRequest("옵션", 123L)));
+        ProductDto dto = new ProductDto("아이스티", 2500, "https://example.com", categoryId, List.of(new OptionDto("옵션", 123L)));
 
         given(productRepository.save(any(Product.class))).willReturn(new Product());
         given(categoryRepository.findById(anyLong())).willReturn(Optional.of(new Category()));
@@ -170,7 +169,7 @@ class ProductServiceTest {
     void addOption() throws Exception {
         //given
         Long productId = 1L;
-        OptionDto dto = new OptionDto("옵션", 2500L);
+        gift.dto.OptionDto dto = new gift.dto.OptionDto("옵션", 2500L);
 
         given(productRepository.findById(anyLong())).willReturn(Optional.of(new Product()));
